nosql_app-master/requirements.txt
attrs==19.3.0
bcrypt==3.1.7
cached-property==1.5.1
cassandra-driver==3.23.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cryptography==2.9.2
dnspython==1.16.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
elasticsearch==7.6.0
eventlet==0.25.2
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-PyMongo==2.3.0
Flask-SocketIO==4.3.0
Flask-WTF==0.14.3
geomet==0.1.2
greenlet==0.4.15
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kafka-python==2.0.1
MarkupSafe==1.1.1
monotonic==1.5
paramiko==2.7.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
PyNaCl==1.3.0
pypiwin32==223
pyrsistent==0.16.0
python-dotenv==0.13.0
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2020.1
pywin32==227
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
six==1.14.0
texttable==1.6.2
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
